Move decode_value and decode_string_value to the printable types crate
