+++ compiler/noirc_printable_type/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 
 [dependencies]
 acvm.workspace = true
+iter-extended.workspace = true
 serde.workspace = true
 
 [dev-dependencies]
