+++ tooling/noirc_abi/src/lib.rs
@@ -12,7 +12,7 @@ use acvm::{
 };
 use errors::AbiError;
 use input_parser::InputValue;
-use iter_extended::{try_btree_map, try_vecmap, vecmap};
+use iter_extended::{try_btree_map, try_vecmap};
 use noirc_printable_type::{PrintableType, PrintableValue, PrintableValueDisplay};
 use serde::{Deserialize, Serialize};
 use std::borrow::Borrow;
@@ -27,10 +27,9 @@ mod arbitrary;
 
 pub mod errors;
 pub mod input_parser;
-mod printable_type;
 mod serialization;
 
-pub use printable_type::decode_value as decode_printable_value;
+pub use noirc_printable_type::{decode_string_value, decode_value as decode_printable_value};
 
 /// A map from the fields in an TOML/JSON file which correspond to some ABI to their values
 pub type InputMap = BTreeMap<String, InputValue>;
@@ -415,18 +414,6 @@ pub fn decode_value(
     };
 
     Ok(value)
-}
-
-pub fn decode_string_value<F: AcirField>(field_elements: &[F]) -> String {
-    let string_as_slice = vecmap(field_elements, |e| {
-        let mut field_as_bytes = e.to_be_bytes();
-        let char_byte = field_as_bytes.pop().unwrap(); // A character in a string is represented by a u8, thus we just want the last byte of the element
-        assert!(field_as_bytes.into_iter().all(|b| b == 0)); // Assert that the rest of the field element's bytes are empty
-        char_byte
-    });
-
-    let final_string = str::from_utf8(&string_as_slice).unwrap();
-    final_string.to_owned()
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
